ORAN-Automation/CI_CD/QA_Testing/requirement.txt
bcrypt==4.0.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==2.0.12
configparser==5.3.0
cryptography==41.0.1
defusedxml==0.7.1
et-xmlfile==1.1.0
fpdf2==2.5.5
httplib2==0.21.0
idna==3.4
ifcfg==0.22
lxml==4.9.2
ncclient==0.6.12
numpy==1.24.3
openpyxl==3.1.2
pandas==1.5.3
paramiko==3.2.0
pexpect==4.8.0
Pillow==9.5.0
ptyprocess==0.7.0
pycparser==2.21
PyNaCl==1.5.0
pyparsing==3.0.9
PyQt5-Qt5==5.15.2
python-dateutil==2.8.2
pytz==2023.3
PyVISA==1.13.0
PyVISA-py==0.7.0
requests==2.27.1
scapy==2.5.0
scp==0.14.5
six==1.16.0
tabulate==0.9.0
typing-extensions==4.6.3
urllib3==1.26.14
xmltodict==0.13.0

